requirements.txt
# Core Video Processing
moviepy>=1.0.3
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Audio Processing
faster-whisper>=0.9.0
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0
# Voiceover: Edge TTS locale (fallback se VELOX_VOICEOVER_TTS_BACKEND non è impostato)
edge-tts>=6.1.0

# Machine Learning / AI
torch>=2.0.0
spacy>=3.7.0

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
gradio>=4.0.0

# Web Automation
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# HTTP & Requests
requests>=2.31.0
Pillow>=10.0.0
urllib3>=2.0.0

# Data Processing
pandas>=2.0.0
rapidfuzz>=3.5.0
dateparser>=1.2.0

# Text Processing & NLP
pysrt>=1.1.2
deep-translator>=1.11.4
argostranslate>=1.9.6

# Google Services
google-auth>=2.23.0
google-api-python-client>=2.100.0
gdown>=4.7.0

# AI APIs
openai>=1.3.0
groq>=0.4.0

# Utilities
pyperclip>=1.8.2
psutil>=5.9.0
validators>=0.22.0

# Visualization (optional but used in dashboard)
plotly>=5.17.0

# Scalability & Cloud Storage (NEW - per 1000+ worker)
boto3>=1.34.0
botocore>=1.34.0
pytrends>=4.9.0

# Database (opzionale - SQLite built-in, PostgreSQL se necessario)
# psycopg2-binary>=2.9.9  # Decommentare se si usa PostgreSQL

# System & File Operations
# (os, sys, json, pathlib, shutil, tempfile, zipfile, subprocess are built-in)
